Initial Commit
Initial commit for the first two modules. This contains a couple of Jupyter notebooks, as well a poetry environment with all of the necessary packages. To start playing with the notebooks, just install poetry and then do "poetry install". Once you have attached the notebook to the generated virtual environment, you should be ready to go. Currently in the middle of doing the project for module 2.