Module 2 - Project - Problem 2 - Part 1
Almost finished the first part of the second problem on the module's project. Its just basic visualization tasks. So I think I have gotten a plot that shows the 3 cell types in the PCA plot (a little bit less self explanatory in the MDS plot) and one that shows 3 clusters and various subtypes in the tsne that uses 1250 PCA components as a basis. Cool stuff