Module 2 - Project - Problem 2 - FInished
Finally finished the last point of the problem number two. It was pretty easy, but running the KMeans method a total of 16 times was very costly, took about 1 hour with my Desktop PC. So what I did now was instead of including the plot for the silhouette score in the same place where the KMeans was run, I split the cells to have one for running KMeans, then one for calculating the silhouette scores and one for plotting the scores. Now if a problem ocurred, I didnt have to re-run the 1hour operation to get my plots back. Now onto part 3!