Module 2 - Project - Problem 3 - Finished!
Finally finished the last (and largest) part of problem 3. I have my doubts about the methodology oI used to discuss the effect of PCs, but I hope its enough to at least get a good chunk of the final score for the homework. Also I included the final rendered report for the homework. Had to basically export each of the jupyter notebooks as HTML documents, open them in the browser and then select "save as PDF" there.