Module 3 - Lecture 1 - Part 4
Initial files for the first lecture of module 3. The first three pages of the lecture are simple videos and material for the course, I included the excercises for part 4. For those I needed to generate plots for graphs, so I included in the poetry installation the networkx package. Seems like a cool solution and the course will probably use it down the road.