Module 3 - Project - Problem 1 and Half of 2
Included the first two problems for the 3rd module project. The first problem was properly finished (it was pretty short) and the second problem is on its way. Also included some notes on how to install the "pygraphviz" package, but it seems like its installation with poetry is much more of a hassle than I thought. I couldnt get it to work.