Module 3 - project - First Three Report Questions
Started to write the report for the module 3 project. Unfortunately I couldnt get to far, but I should finish the rest pretty quick. Also, decided to install an "Export to PDF" extension in VSCode that allows me to export a Markdown file into a PDF. It was mainly done to write a simple markdown document as the report and then have it be rendered as a PDF when the time comes to submit it.