Module 3 - Final Written Report
FInished the written report for module 3. I had to rush the last part, but I managed to deliver it on time. I includes a couple of new images and sections in the script, since it was used for the last part that I hadn't finished, the project that you could create with whathever data you wanted.