Module 4 - Recitation Notebooks
Included the two Jupyter notebooks for the module 4 recitation. I started going through the climate signal and making it compatible with the dependencies found in my pyproject.toml file in the repo. Seems like the notebook is using a deprecated version of seaborn that doesn't run completely seamlessly with the latest version. Started making some changes.