Module 4 - Initial Project Files
Included the initial files for the project of the fourth module, this includes both the jupyter notebooks for problems 1 and 2, and the written report markdown file to generate the report for next week. Also decided to include the datasets for the project, since they are freely available online as well.