Module 4 - Project - Problem 3
Almost finished the autograded section of the project. The last problem (relating to a BER calculation thats meant to convert the BER metrics into percentages as a way to later check if the BER has any impact on the CPI) was left finished, but incorrect, as I didnt know how to exactly complete it.