Module 4 - Project - Written Report for Problems 1 and 2
Finished the first 2 problems of the written report. The first one was really except for the last question. where I had to look at the forum below to see if I could get my questions answered. After that the second problem was very informative, as I actually didnt really understand how to calculate the autocovariance function of an AR or MA model. Now I actually get it.