Module 5 - Recitation Files and Initial Project Notebook
Included the files for the two recitations in the module (both the Matlab livescripts and Jupyter notebooks) and also included a simple notebook for the autograded section of the project. Decided to ignore the data for the project, since its too large to push and could compromise the privacy of the course.