Module 5 - Project: Simulation Tests
Finished the part where it asks us to simulate the movement of a set of particles across the Phillipinne archipielago. Now that its working, I went ahead and checked future sections and realized that the simulation algorithm will be used in multiple parts, so I will create a set of helper functions to reuse the logic I created across multiple points.