Module 5 - Project: Problem 4
Finished all of the points found in problem 4. Also managed to contain the logic for optimizing the kernel params (gaussian process), the different kernels available, and the plots to compare the grid search log likelihood in functions. That way the logic used throughout the notebook can be very easily repurposed for new problems in the future. Next step actually, is refactoring the logic for calculating the conditional estimate of the variance and mean.

Other changes:
- Moved the flow simulation gif to its own folder