Module 5 - Project: Problem 5?
Finished the prediction of new data. However, something seems to be going horribly wrong, as my predictions are going way overboard and dont fit to the actual data in any way. However, I finished the prediction routine in order to facilitate its use. Also removed the manual calculation of the kernel inside of the optimization routine, and moved said routine to using the "predict_conditional_mean_and_variance" function. Also included a new fix that now transforms the values of X provided to the function into indices that it can use to index sigma freely, instead of using X directly and just assuming that X should consist of indices.