Module 5 - Project: Kernel and Log Likelihood Fixes
I was getting very bad prediction results on one of the last points of the project. Found a couple of bugs and fixed them. This gave me results that are a bit better, but in the tails of the prediction they start going crazy. These are some of the fixes:
- Removed the square from the sigma variable, and added a 2 dividing the difference in measurements. Not sure if this is doing anything, but at least it caused the values of the kernel to not go so low, as the sigma scaling factor is not squared.
- The ideal set of parameters is the one that generates the highest log likelihood. Previously I was fetching the lowest value thinking it was a traditional optimization problem.

Other changes:
- Fixed the "y" range that is allowed to be plotted in the final prediction plot. Since the last set of predictions go wild in the tail end, I simply limited the values that should be plotted.
- Fixed some Latex in the recitation notebook.