Module 5 - Project: RBF Kernel and Plots
Decided to overhaul the RBF kernel and generate all of the plots in the notebook with with said kernel. I added logic to automatically download the generated plots to both the "gaussian_process.py" file and the "Project - Part II.ipynb" jupyter notebook. Also fixed some changes in the recitation notebook in terms of its LateX rendering. The only unfortunate thing about the RBF kernel is that its predictions on the tail end start going crazy. In the next commit I will add support for different kernels so that I can try the rational quadratic kernel (or any other that might be necessary in case that one doesn't work).