Module 5 - Project: Adapted Logic to Support Kernels with N Params
Adapted the functions and the main notebook, to now allow me to input any kernel function, and then simply change the ranges to test. Everything adapts automatically to the growing number of parameters. The only things that dont adapt are the plots, since I didnt want to do 3D plots. I simply limited the plots to 2D and everything above doesn't get plotted. The generated images changed a bit after changing the kernel.