Module 5 - Project: New Kernel and Mean Method
The RBF method was meh at fitting the data, as the trailing section of the predictions that it generated were going off into crazy territorry. So what I did is take into account some of the suggestions from the course: Change the method used to calculate the prior of the x1 mean and the kernel used. The kernel improved things... a bit, but not in a particularly big way. So what I did was change the method to calculate the mean from a moving average, to regular old mean of all the data and maaan did that make a difference. Now the results actually look correct. Its great. With this new data I will generate the last two simulation points and get working on the report (should be a piece of cake with everything that I have).

Other changes:
- Solved a bug where the grid search plot didnt show up correctly due to weird indexing.