Module 5 - Project: Finished Mandatory Part of the Report
Managed to finish the mandatory section of the report. The first section, where the correlation of points is used to find dipoles, was skipped in order to make way for the rest of the report. Also had to add a couple of new sections to the part II notebook, that create a progression of the final simulation and a big plot that was modified with photoshop to add the possible location of outposts. Another important change, was that when a larger number of particles, the simulation algorithm tended to fail. This was due to the surrogate locations used to index the v_t array, were going over the available indices in v_t. After the fix, everything ran smoothly.